pyproject.toml
[project]
name = "tua-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "geopandas>=1.1.3",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "matplotlib>=3.8.0",
    "onnx>=1.21.0",
    "onnxruntime>=1.24.4",
    "opencv-python>=4.13.0.92",
    "pycbc>=2.10.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "python-multipart>=0.0.22",
    "rasterio>=1.5.0",
    "scipy>=1.12.0",
    "ultralytics>=8.4.30",
    "uvicorn>=0.42.0",
]
